+++ src/SnippetPredictor.Test/Library.fs
@@ -351,8 +351,12 @@ module SnippetPredictorInitialization =
             [
 
               test "run" {
+                  use tmp = new TempFile(".snippet-predictor.json", """{"Snippets": []}""")
+                  use _ = new EnvironmentVariable(tmp.GetSnippetDirectoryPath())
                   let subsystem = Init()
                   (subsystem :> IModuleAssemblyInitializer).OnImport()
+
+                  Async.Sleep(1000) |> Async.RunSynchronously
 
                   let predictor = getSnippetPredictorSubsystem ()
                   let predictor = predictor |> Expect.wantSome "should have Snippet predictor"
@@ -369,6 +373,9 @@ module SnippetPredictorInitialization =
                   Integration.getExactIdentifierSnippetTexts ":unsupported"
                   |> Expect.isEmpty "should use the registered predictor integration"
 
+                  Integration.isUnknownGroupIdentifier ":unsupported"
+                  |> Expect.isTrue "should use the registered predictor integration"
+
                   (subsystem :> IModuleAssemblyCleanup).OnRemove(createMockModule ())
 
                   let predictor = getSnippetPredictorSubsystem ()
@@ -379,6 +386,9 @@ module SnippetPredictorInitialization =
 
                   Integration.getExactIdentifierSnippetTexts ":snp"
                   |> Expect.isEmpty "should clear the predictor integration"
+
+                  Integration.isUnknownGroupIdentifier ":unsupported"
+                  |> Expect.isFalse "should clear the predictor integration"
               }
 
               ]
@@ -505,6 +515,22 @@ module SnippetPredictor =
 
                   predictor.GetExactIdentifierSnippetTexts(":group Echo")
                   |> Expect.isEmpty "should exclude snippet search input"
+              }
+
+              test "IsUnknownGroupIdentifier" {
+                  use predictor =
+                      new SnippetPredictorForTest(testAssetPath ".snippet-predictor-valid.json")
+
+                  Async.Sleep(1000) |> Async.RunSynchronously
+
+                  predictor.IsUnknownGroupIdentifier(":unknown")
+                  |> Expect.isTrue "should identify an unknown group identifier"
+
+                  predictor.IsUnknownGroupIdentifier(":group")
+                  |> Expect.isFalse "should identify a configured group identifier"
+
+                  predictor.IsUnknownGroupIdentifier(":group Echo")
+                  |> Expect.isFalse "should exclude group-scoped input"
               }
 
               test "for coverage" {